Various Additions and Changes
-Added setusersound command.
-Added showusersounds command.
-Added the ability to escape bot name keyword response in a message by
preceeding at least one with a backslash ("\").

-Updated conditions when a user sound would play;
Now it only plays when a user connects to a voice channel on the server
(not including moving channels) that the bot is allowed to play in, if
the bot is not curently present in any channel, and there are other
people in the voice channel.

-Fixed crashing error with sound playback.