Added configuration templates for StartupCheck to reset to
Added Templates folder, containing default_config.json and
default_permissions.json. The contents of these files are used to reset
the configuration files if StartupCheck determins there is an issue with
them.
This will also allow changes to be made more easily to the configuration
files in future updates/changes.