Implemented ability for admins to modify audio tags/artists
songinfo and soundinfo commands will now parse additional arguments to
modify data for the requested sounds. If arguments start with specific
characters, the argument will be added/removed as a tag/artist for the
sound/song. Underscores will be replaced with spaces.
"+&" adds the argument as an artist (case-sensitive).
"-&" removes the argument from artists (case-sensitive).
"+" adds the argument as a tag.
"-" removes the argument from tags.